chore(mcp): extract shared chart helpers and ASCII rendering into separate modules
- Create chart_helpers.py with find_chart_by_identifier, get_cached_form_data,
  extract_form_data_key_from_url (deduplicated from 7+ tool files)
- Extract ~866 lines of ASCII chart rendering from get_chart_preview.py (2283 lines)
  into ascii_charts.py, modernize type hints to Python 3.10+ style
- Replace broad except Exception catches with specific exceptions in 4 chart tools
- Fix fragile str.split URL parsing in update_chart_preview.py
- Add unit tests for chart_helpers.py

[sc-104270]
