+++ superset/mcp_service/chart/tool/get_chart_info.py
@@ -26,8 +26,8 @@ from sqlalchemy.orm import subqueryload
 from superset_core.mcp.decorators import tool, ToolAnnotations
 
 from superset.commands.exceptions import CommandException
-from superset.commands.explore.form_data.parameters import CommandParameters
 from superset.extensions import event_logger
+from superset.mcp_service.chart.chart_helpers import get_cached_form_data
 from superset.mcp_service.chart.chart_utils import validate_chart_dataset
 from superset.mcp_service.chart.schemas import (
     ChartError,
@@ -40,26 +40,11 @@ from superset.mcp_service.mcp_core import ModelGetInfoCore
 logger = logging.getLogger(__name__)
 
 
-def _get_cached_form_data(form_data_key: str) -> str | None:
-    """Retrieve form_data from cache using form_data_key.
-
-    Returns the JSON string of form_data if found, None otherwise.
-    """
-    from superset.commands.explore.form_data.get import GetFormDataCommand
-
-    try:
-        cmd_params = CommandParameters(key=form_data_key)
-        return GetFormDataCommand(cmd_params).run()
-    except (KeyError, ValueError, CommandException) as e:
-        logger.warning("Failed to retrieve form_data from cache: %s", e)
-        return None
-
-
 def _build_unsaved_chart_info(form_data_key: str) -> ChartInfo | ChartError:
     """Build a ChartInfo from cached form_data when no chart identifier exists."""
     from superset.utils import json as utils_json
 
-    cached_form_data = _get_cached_form_data(form_data_key)
+    cached_form_data = get_cached_form_data(form_data_key)
     if not cached_form_data:
         return ChartError(
             error="No cached chart data found for form_data_key. "
@@ -92,7 +77,7 @@ def _apply_unsaved_state_override(result: ChartInfo, form_data_key: str) -> None
     """Override a ChartInfo's form_data with cached unsaved state."""
     from superset.utils import json as utils_json
 
-    if cached_form_data := _get_cached_form_data(form_data_key):
+    if cached_form_data := get_cached_form_data(form_data_key):
         try:
             result.form_data = utils_json.loads(cached_form_data)
             result.form_data_key = form_data_key
