+++ superset/mcp_service/chart/tool/update_chart_preview.py
@@ -26,8 +26,12 @@ from typing import Any, Dict
 from fastmcp import Context
 from superset_core.mcp.decorators import tool, ToolAnnotations
 
-from superset.exceptions import OAuth2Error, OAuth2RedirectError
+from sqlalchemy.exc import SQLAlchemyError
+
+from superset.commands.exceptions import CommandException
+from superset.exceptions import OAuth2Error, OAuth2RedirectError, SupersetException
 from superset.extensions import event_logger
+from superset.mcp_service.chart.chart_helpers import extract_form_data_key_from_url
 from superset.mcp_service.chart.chart_utils import (
     analyze_chart_capabilities,
     analyze_chart_semantics,
@@ -123,9 +127,7 @@ def update_chart_preview(
             explore_url = generate_explore_link(request.dataset_id, new_form_data)
 
         # Extract new form_data_key from the explore URL
-        new_form_data_key = None
-        if "form_data_key=" in explore_url:
-            new_form_data_key = explore_url.split("form_data_key=")[1].split("&")[0]
+        new_form_data_key = extract_form_data_key_from_url(explore_url)
 
         with event_logger.log_context(action="mcp.update_chart_preview.metadata"):
             # Generate semantic analysis
@@ -199,7 +201,15 @@ def update_chart_preview(
             "error": OAUTH2_CONFIG_ERROR_MESSAGE,
             "success": False,
         }
-    except Exception as e:
+    except (
+        SupersetException,
+        CommandException,
+        SQLAlchemyError,
+        KeyError,
+        ValueError,
+        TypeError,
+        AttributeError,
+    ) as e:
         execution_time = int((time.time() - start_time) * 1000)
         return {
             "chart": None,
